Move everything related to tvb-simulator in its own package
