+++ tvbwidgets/tests/test_tvb_integrators.py
@@ -1,5 +1,11 @@
-import pytest
-from tvbwidgets.core.tvb_integrators import IntegratorsEnum
+# -*- coding: utf-8 -*-
+#
+# "TheVirtualBrain - Widgets" package
+#
+# (c) 2022-2023, TVB Widgets Team
+#
+
+from tvbwidgets.core.simulator.tvb_integrators import IntegratorsEnum
 
 
 def test_get_integrators_dict():
